Remove unused variables from Atmos linear model
The size of the linear systems built from the Atmos linear model
directly depend on the number of state variables.  For example, this
becomes a performance bottleneck as the number of tracers is increased
in a GCM configuration with 1D IMEX.  Thus, we remove the unused
variables from the linear model.