Allow multiple numerical fluxes in remainder model
By allowing the remainder model to take a tuple of fluxes, the remainder
can become a true discrete splitting, meaning that the sum of remainder
model and its subcomponents is the same of the main model

The default for both the driver and the remainder model is a physics
level splitting not a discrete splitting. This is because some testing
has revealed that this might be more numerically stable.