Allow multiple direction in remainder model
By allowing multiple directions in the remainder model one can, for
example, define a vertical model then the remainder as the full model
minus the vertical. An example of the use of this is doing implicit time
stepping for the vertical acoustic waves, and explicit time stepping for
the rest of the model.